implemented exercise features
Creating offline mode with a message and a cache so that all messages are still shown.
the user can clear this cache to delete messages locally.
Disabled the input message text field for offline mode.
adjusted CSS a bit for the container box on the main screen, got rid of the message on the top and fixed the 'Send button'